refactor: retarget versioning owner
